// src/app/main/limt-quota/constants/limit-quota.enums.ts
export enum LIMITANDQUOTA {
    COMPANY = 'api/referencedata/company',
    GENDER = 'gender',
    CATEGORY = 'category',
    BRAND_TYPE = 'brandType',
    PERIOD = 'api/referencedata/period'
}

export enum LimitAndQuotaTablePath {
    COMPANY = 'api/limits/company',
    CATEGORY = 'api/limits/category',
    CATEGORYHEADER = 'api/limits/header/category',
    COMPANYHEADER = 'api/limits/header/company',
    DOWNLOAD_COMPANY = 'api/limits/download/company',
    DOWNLOAD_CATEGORY = 'api/limits/download/category',
}